move FileCompressionType out common and into core
