+++ datafusion/common/src/lib.rs
@@ -46,9 +46,8 @@ pub use error::{
 };
 
 pub use file_options::file_type::{
-    FileCompressionType, FileType, GetExt, DEFAULT_ARROW_EXTENSION,
-    DEFAULT_AVRO_EXTENSION, DEFAULT_CSV_EXTENSION, DEFAULT_JSON_EXTENSION,
-    DEFAULT_PARQUET_EXTENSION,
+    FileType, GetExt, DEFAULT_ARROW_EXTENSION, DEFAULT_AVRO_EXTENSION,
+    DEFAULT_CSV_EXTENSION, DEFAULT_JSON_EXTENSION, DEFAULT_PARQUET_EXTENSION,
 };
 pub use file_options::FileTypeWriterOptions;
 pub use functional_dependencies::{
