+++ datafusion/core/src/physical_optimizer/replace_with_order_preserving_variants.rs
@@ -281,11 +281,11 @@ mod tests {
 
     use crate::prelude::SessionConfig;
 
+    use crate::datasource::file_format::file_compression_type::FileCompressionType;
     use crate::datasource::listing::PartitionedFile;
     use crate::datasource::physical_plan::{CsvExec, FileScanConfig};
     use crate::physical_plan::coalesce_batches::CoalesceBatchesExec;
     use crate::physical_plan::coalesce_partitions::CoalescePartitionsExec;
-    use datafusion_common::FileCompressionType;
 
     use crate::physical_plan::filter::FilterExec;
     use crate::physical_plan::joins::{HashJoinExec, PartitionMode};
